git-svn-id: http://svn.berlios.de/svnroot/repos/sems/tags/rel_0.10.0@498 8eb893ce-cfd4-0310-b710-fb5ebe64c474

+++ apps/reg_agent/RegistrationAgent.cpp
@@ -180,7 +180,7 @@ void RegThread::run() {
 	 it != registrations.end(); it++) {
       if (!check_registration(*it)) {
 	DBG("Registration %d does not exist or timeout. Creating registration.\n",
-	    it - registrations.begin());
+	    (int)(it - registrations.begin()));
 	create_registration(*it);
       }
     }
